requirements.txt
PyOpenGL==3.1.4
PyQt5==5.13.2
PyQt5-sip==12.7.0
